# docs/swagger.yaml
definitions:
  entity.BusinessLoginOrRegisterRequest:
    properties:
      code:
        description: 验证码
        type: string
      tel:
        description: 手机号码
        type: string
    type: object
  entity.BusinessPassLoginRequest:
    properties:
      password:
        description: 密码
        type: string
      tel:
        description: 手机号码
        type: string
    type: object
  entity.ResponseData:
    properties:
      data:
        additionalProperties: true
        description: 返回数据
        type: object
      message:
        description: 提示信息
        type: string
      status:
        description: 成功失败标志；true：成功 、false：失败
        type: boolean
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /api/v1/business/codelogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.BusinessLoginOrRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/entity.ResponseData'
      summary: 商家手机验证码登录
      tags:
      - 商家
  /api/v1/business/passlogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.BusinessPassLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/entity.ResponseData'
      summary: 商家账号密码登录
      tags:
      - 商家
  /api/v1/business/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.BusinessLoginOrRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/entity.ResponseData'
      summary: 商家注册
      tags:
      - 商家
  /api/v1/business/store:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/entity.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 查询商家门店信息
      tags:
      - 商家
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: 门店名称
        in: formData
        name: store_name
        required: true
        type: string
      - description: 门店详细地址
        in: formData
        name: store_address
        required: true
        type: string
      - description: 门店logo
        in: formData
        name: store_logo
        type: file
      - description: 门店联系人姓名
        in: formData
        name: store_contact_name
        type: string
      - description: 门店联系人电话
        in: formData
        name: store_contact_tel
        type: string
      - description: 门店开始营业时间
        in: formData
        name: store_start_banking_hours
        required: true
        type: string
      - description: 门店结束营业时间
        in: formData
        name: store_end_banking_hours
        required: true
        type: string
      - description: 门脸照
        in: formData
        name: store_face_photo
        type: file
      - description: 店内照
        in: formData
        name: in_store_photos
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/entity.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新商家门店信息
      tags:
      - 商家
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"